Started the general code cleanup

+++ shaders/lib/util/jitter.glsl
@@ -1,5 +1,5 @@
 //Jitter offset from Chocapic13
-uniform float framemod8;
+uniform int framemod8;
 
 vec2 jitterOffsets[8] = vec2[8](
 							vec2( 0.125,-0.375),
@@ -13,6 +13,6 @@ vec2 jitterOffsets[8] = vec2[8](
 						);
 							   
 vec2 TAAJitter(vec2 coord, float w) {
-	vec2 offset = jitterOffsets[int(framemod8)] * (w / vec2(viewWidth, viewHeight));
+	vec2 offset = jitterOffsets[framemod8] * (w / vec2(viewWidth, viewHeight));
 	return coord + offset;
 }
